+++ .gitignore
@@ -3,3 +3,4 @@
 /build/
 /dist/
 /bin
+/nbproject/pmd.settings
