+++ src/net/erbros/Lottery/LotteryDraw.java
@@ -1,0 +1,24 @@
+package net.erbros.Lottery;
+
+import java.util.TimerTask;
+
+    class LotteryDraw extends TimerTask {
+        
+    final private Lottery plugin;
+    final private boolean draw;
+
+    public LotteryDraw(final Lottery plugin, final boolean draw) {
+        this.plugin = plugin;
+        this.draw = draw;
+    }
+
+        public void run() {
+
+            if (draw && plugin.isLotteryDue()) {
+                plugin.lotteryDraw();
+            }
+            else {
+                plugin.extendLotteryDraw();
+            }
+        }
+    }
